% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_surfaceAICPlot.R
\name{extended_surfaceAICPlot}
\alias{extended_surfaceAICPlot}
\title{Plot AICc values of multi-OU models}
\usage{
extended_surfaceAICPlot(ext_surface, summary, fwd_surface = NA)
}
\arguments{
\item{ext_surface}{List of models obtained using the extended phase of
SURFACE. This is always the first element in the list returned by
\code{surfaceExtended}.}

\item{summary}{A data.frame with AICc values of OUwie models. This is always
the second element in the list returned by \code{surfaceExtended}.}

\item{fwd_surface}{Optional. List of models obtained using the forward phase
of SURFACE. If provided these will also be included in the plot.}
}
\value{
Plot of AICc values against number of regimes.
}
\description{
Explore model fit by visually comparing the AICc values of the results
obtained using \code{surfaceExtended}. The fit of OUM models obtained using
forward, backwards and extended phases of SURFACE, as well as that of models
explored with OUwie, are plotted against the number of regimes.
}
\seealso{
\code{\link{surfaceExtended}}
}
