% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions_through_time.R
\name{transitions_through_time}
\alias{transitions_through_time}
\title{Summarize trait dynamics through time}
\usage{
transitions_through_time(tree, char, repl = 100, model = "ER", plot = T)
}
\arguments{
\item{tree}{Phylogenetic tree in 'phylo' format.}

\item{char}{Named vector including character states for all tips. Names need
to correspond to tips in the phylogeny.}

\item{repl}{Number of replicates of stochastic character mapping (default = 100)}

\item{model}{A character specifying the model of evolution used to
reconstruct the evolutionary history of the character. Options include
\code{'ER'} (default), \code{'SYM'} and \code{'ARD'}, see
\code{\link[ape]{ace}}.}

\item{plot}{A logical indicating whether to plot the results. Default is
\code{TRUE}.}
}
\value{
A data.frame including the times of birth and death of each instance
  of a character across replicates. This object can be passed to
  \code{\link{plot_ttt}} to obtain different visualizations.
}
\description{
This function provides a way of summarizing the temporal dynamics of traits by
registering the times when there are transitions between character states.
First, a number of replicates of stochastic character mappings are created
(see \code{\link[phytools]{make.simmap}} for more details), and the phylogeny
is traversed in order to register the moments in time when there are
transitions between states. For every time a transition occurs, the function
will follow all descendants and also register the point in time where that
instance of a given state goes extinct (either by all descendants
transitioning to a different state or by all tips going extinct).
}
\details{
The funtion only requires a tree and a character, and allows the user to
modify the number of replicates, the model used for character mapping, as well
as whether to return a plot summarizing the results. More control on what and
how to plot these results is attained by providing the output to the function
\code{\link{plot_ttt}}.
}
\references{
Mongiardino Koch N., Thompson J.R. A Total-Evidence Dated Phylogeny of Echinoids and the Evolution of Body Size across Adaptive Landscape. bioRxiv 2020.02.13.947796, doi.org/10.1101/2020.02.13.947796.
}
\seealso{
For details on stochastic character mapping visit
  \code{\link[phytools]{make.simmap}}. Plots can be explored with
  \code{\link{plot_ttt}}
}
\author{
Nicolás Mongiardino Koch
}
